Shim egress operations
